SCBTruncate can select sub-set of agents
scbTruncate has added some arguments which allow for a subset of agents
as well as a subset of frames.  Currently, it simply takes the first N
agents from the scene.  Ultimately, it would be nice if it could randomly
select them.  That way, even if the definitions are spatially related,
the sampling could be over the entire space.