Added proper rasterization
Density (aka position) rasterization has used an approximation.  The
position of each agent is constrained to the nearest cell center.  This
makes it so that every agent's effect on its neighborhood is the same.

The "proper" thing to do is to eliminate that approximation and have
each agent affect it's neighborhood based on the real difference.  I added
this functionality (using Kernel2 and rasterizePosition2).  It's MUCH
slower (6-7X more time for 25K agents.)  For a cell size of 0.2, it's
not significantly different.