Speed rasterized discretely
Having convinced myself that a speed field isn't very meaningful, I've
changed the speed rasterizer so that it simply rasterizes the agent, where
it is with it's speed.  Last agent on a cell wins.

Added some additional code so that instead of single pixels, it rasterizes
to a 3x3 square of pixels.  Makes things more visible.