Obstacle parser made a bit more robust
I can now feed full experiments into it and it'll skip agent sets and
the like, focusing purely on the obstacles.

This was necessary because Goals (in AgentSet) also use the Vertex tag.
I needed to make sure I actually was defining an obstacle when I hit a vertex
definition.