Field smoothing tool
The functionality for smoothing a vector field.  NOTE: the smoothing in
no way preserves vector magnitude.  Typically, if the magnitudes point in
different directions, the magnitudes will gradually shrink.  It's important
to periodically renormalize.

I might make an option/version which preserves magnitude by automatically
rescaling the resultant vectors to the original magnitudes.

The good news is it looks like it worked.  The SMOOTHED tangent field is
better behaved.