Increased the domain of the tawaf grid
Correspondingly, I removed the mask blitting that I was doing because the
new grid doesn't align with the mask anymore.  I should do the mask in a
more intelligent way (such as drawing the mask polygon transparently.)
In that case, it would scale arbitrarily.  I currently do that with the
regions.