Remvoed advec flow analysis
I had an old analysis technique in which you could advect a few agents
specified by a line.  It was a piece of crap analysis.

I've replaced it with legitimate flow analysis.  Draw lines and then count
the people moving across it.  Currently, I only have the GUI end of it
working.  I need to connect it to the actual analysis.