Factored config out
The configuration class (Config) has been removed from analyze crowd.
I'm going to use the same file used by the gui as the mechanism for
specifying arguments to the consistency.  Ultimately, I'll use it for the
others as well.  That way, the work object doesn't need to know as much
about each function, it can just pass the configuration and allow it to
complain if the information is insufficient.