NPFrameSet made compatible with Julich Trajectories
Julich trajectories have can have a different population from frame to frame.
In order for both objects to work in the same context, I need to have info
in each frame of WHO the agents in the frame are.  The Julich data provides
global ids for each of the frame ids.

For the NPFrameSet, the frame ids ARE the global IDs, so I simply give a
class which reflects the frame index back.