Changed kernel discretization code
I changed how the code selects the number of cells the kernel requires.
Previously, if the support area was an integer multiple of the cell size,
it may not actually produce that integer multiple because of floating point
error (the ceiling would take any error on the high side and immediately
force a new cell.)  In this case, fixboundary would be called and that
function isn't necessarily smart enough to handle those infinitesimal
overlaps.

Instead, I test for a threshold.  Works much better.