Fixed grid initialization
I had a problem in which I didn't intialize the convolution grid.  For
convolution with a field signal, this was the right thing to do.  HOwever,
convolution with dirac signal needs a properly initialized grid.  So, I
added a function on the kernel which looks at the signal and reports if
initialization is necessary and what value should be used.