Fixed the bug in GFSSignal
There was a bug in GFSSignal which was causing every thread to repeat the
same results over and over.  I was trying to copy a DataGrid into a numpy
array which was not supported.

In the ThreadRasterization, this was stomped on by the exception handling
and it would silently fail, using the previous grid.  This is no longer
the case.