Dirac convolution can be done fast or precise
Added an additional flag to the kernel to specify what type of dirac
convolution is performed: fast or precise.  In the precise convolution,
the distances from the dirac signal to the cell CENTERS is computed and
the kernel function is evaluated on those values.  Thus, for each
dirac function, a unique kernel is computed.  In the fast version, the
cell center, in which the dirac signal lies, is used.  In this case, the
same, pre-computed kernel can be used over and over.