Removed the illusion of precision
When the kernel's support area isn't perfectly aligned with the grid then
some inaccuracies creep in.  I had some code that somewhat pretended to
deal with it, but, in reality, it didn't.  So, I've stripped out all
that code in favor of an even, approximate result straight across the
board.

The most "correct" thing to do (as I consider it) is as follows.  Each cell
should contain the integral of the function over that cell, divided by the
cell area.  If I were to actually do this, the integral of the full kernel
would be one.  Anything short of that is wrong.  So, if I'm going to be
wrong, I should be simple.

1) Removed "diracFast" and "diracPrecise".  It's all just fast.
2) Removed the fix 1d boundary.