Kernels changed to handle memory constraints
The kernel can be arbitrarily large (based on support size and cell size)
However, I can't create arbitrarily large arrays.  Solving this so that
it can work for arbitrary sizes is non-trivial.  The most general solution
is INCREDIBLY expensive.

The current "solution" is to put hard-coded limits suitable for the domain
in which I'm applying this functionality.  I do tests and throw exceptions.
IN addition, I've bounded the distance values for the Plaue11Kernel.
The maximum distance prevents unboundedly large kernels.