Updated PlaueKernel so it is memory safe
Because Plaue kernels can get arbitrarily large, it is possible to run out
of memory when pre-computing.  This is particularly ironic because the
majority of the work performed for such large kernels is wasted as the
vast majority of the sample values fall outside the convolution domain.
And, in the light of the fact that the kernel is computed for each agent
from scratch, there is no value in pre-computing.

I've changed the algorithm to dismiss all pre-computing and to perform
only that work that applies to the convolution domain -- thus the amount
of memory required is bounded by the size of the convolution domain.
This is memory safe and thread safe.