GridFileSequence: splatAgents and computeSpeeds updated
Both now use the current infrastructure.
The computeSpeeds is kind of crappy.  I'd be better off creating a new
signal type that uses the pedestrian data to compute speeds at each frame
and then simply convolve a kernel with it.  Although, this approach would
have to be carefully done to be thread safe.