Add the ability to transform trajectories
In order to shift/realign data, I can operate directly on the trajectories
to transform them.  Currently, I only support translation in x and y.
I added a new SCBData class in order to facilitate this.  Unlike FrameSet
and NPFrameSet which read the data on the fly from the hard-drive, the
new hold the entire data set in memory.  I should use this for other
applications in which I examine the size of the data and determine if it
is worthwhile.