Add ability to label flow lines in GUI
Previously, the legends in the flow plots had meaningless names:
Line 0, Line 1, etc.  Now line names can be given in the GUI, they are
included in the computation and the visualization.  Even when data has
already been computed, the names for the plot can be dynamically changed
without recomputing just by changing the name in the gui and selecting
visualize only.  The new name will replace the names stored in the file.