Initial functionality for creating sets of work
I've got a collapsable group box and a layout that handles it reasonbly
intelligently.  Furthermore, the controls are set up in a scrollable
window that handles changes in length due to collapsing boxes.

I also reorganized the window so that the controls span the entire height
of the app, with the gui and console on top of each other.