Update objToNavMesh -- better feedback (#6)
This functionality has historically been brittle. It had a bunch of
implicit assumptions that, if not met, led to inscrutable errors. This adds
explicit integrity checks with meaningful error messages and unit tests on
those integrity tests.

This also allows the user to specify which direction is "up". Previously,
it was hard-coded to assume a y-up world. Now it can be specified on the
command-line. This new functionality is *not* tested.