RoadmapBuilder - Fix field smoothing
This fixes a bug with field smoothing; when the smoothing brush got too
close to the field boundary, it would crash. Too close is defined by
brush size and kernel size. It would have to draw data from a region that
may have extended beyond the field.

This modifies the VectorField.subRegion() method to return empty fields
when accessing regions beyond the valid domains (as represented by cell
indices). That means the smoothing will include lots of zeros near the
boundary.

This also introduces some rudimentary unit tests for both vector field and
field tools. There is still a *ton* to unit test.