Adding unit tests for primitives
This adds unit tests for the classes and functions defined in
primitives.py. It corrects any bugs that are found during the process. The
goal is for tests that pass in both python 2.7 and python 3.9.

This also unifies the APIs between V2 and V3 with one exception; it doesn't
reconcile the length/magnitude and lengthSquare/magSq in V3 (whereas V2
only has the magnitude variant). They should all use a single abstraction
for those quantities. But I'll defer that to clean up.

This *doesn't* clean up the style of the file.